PipelineRunsFilter can have multiple filters (#4425)
I don't think this comment is accurate - or at least completely
accurate. I noticed a test elsewhere that was successfully filtering on
both tags and statuses which seems to violate the documentation. I've
added a test to prove it.

Perhaps what the original description is trying to say is that we use
the intersection of filters? That you can't represent something like
"pipeline runs that are finished or have a certain tag"? I'm not
entirely sure.