[demo pipelines] only write comments and stories tables to snowflake #4659
The demo pipelines do a weird thing where they write the comments and stories to both Snowflake and S3. While this is not something that nobody would ever want to do, it's a weird thing to do, and it causes awkwardness in our implementation that could throw off casual readers.