Always pass through executable path to ReconstructablePipeline (#5458)
Summary:
We missed a path where it was possible to lose track of the exeuctable path when launching a run and switch back to sys.executable. This diff ensures that every time we construct a ReconstructablePipeline from a PipelinePythonOrigin (e.g. in the run worker), it passes through the executable path if executable.

This allows us to bring back the codepath that - but I think we probably want to get a test up and running that explicitly checks that an executable path that is different than sys.executable is actually passed all the way through.

Test Plan: Canary test now passes